@@ -4,10 +4,8 @@ use Addwiki\Mediawiki\Commands\EditPage;
 use Addwiki\Mediawiki\Commands\Purge;
 use Addwiki\Mediawiki\Commands\RestoreRevisions;
 
-$GLOBALS['awwCommands'][] = function ( $awwConfig ) {
-	return [
+$GLOBALS['awwCommands'][] = fn( $awwConfig ) => [
 		new EditPage( $awwConfig ),
 		new Purge( $awwConfig ),
 		new RestoreRevisions( $awwConfig ),
 	];
-};
