@@ -16,7 +16,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class Purge extends Command {
 
-	private $appConfig;
+	private ArrayAccess $appConfig;
 
 	public function __construct( ArrayAccess $appConfig ) {
 		$this->appConfig = $appConfig;
@@ -78,6 +78,7 @@ class Purge extends Command {
 		}
 
 		$output->writeln( 'Done' );
+		return 0;
 	}
 
 }
